feat: add .yamlConfig option
Now it's possible to specify xremap config in yaml format in the new
.yamlConfig module option.

The preexisting .config option is turned into the config file by taking
the nix attset and saving it as a JSON file. This has worked fine for
the time being, but YAML and JSON are not really 1:1. For example it's
possible to embed objects in YAML which can be useful for xremap's
config syntax.

This feature was introduced to address #19 but during testing it was
discovered that it was caused by unnecessary braces in xremap's original
README.

During testing it was discovered that launch: settings apparently do not
work on development nixosConfigurations. In my test rounds the launched
programs were constantly dumping cores.

Too restrictive systemd settings are the prime suspect and will be
investigated in the next commits.