feat: add home-manager module
This commit adds a new output to the flake called "homeManagerModules".
This output contains a home manager module that configures xremap on a
non-NixOS system which is managed through home-manager module.

`nix flake check` marks homeManagerModules as an unknown output since
it's not standard, but some projects (e.g. Hyprland) are using this
schema.

Should a common naming approach be codified -- homeManagerModules output
will be deprecated.

Closes #20