Added link to Systemwide example with X feature for application-specific remaps
This is because this way, it's easier to find the information you are looking for. Without this link, it can be confusing if you mean an example within this repository, or one of the examples in the main repository (https://github.com/xremap/xremap/).

This way, users can immediately find the configuration you're referring to by following the link.