feat: enable Niri support, expose as an option
This commit adds support for Niri WM that was added in xremap 0.10.15.

I was not able to write a proper non-interactive test for niri because
of the limited graphics options in nixosTest environment, but I left a
working interactive test in `.dev/` directory. It will not be a part of
CI until it can be properly run in CI.