refactor: start move to a development subflake
This commit kicks off a set of changes that will ultimately remove
unneeded inputs from the main flake.

Certain inputs (hyprland, home-manager, devshell at the very least) are
not interesting to the flake consumers as they are only used during
development process.

This change starts a development-only subflake that will use all those
inputs to:

1. Show xremap demos
2. Run CI using inputs (now) not in the main flake

This specific commit starts the development flake and adds a demo of
xremap running in hyprland

Related: #84