refactor: move home manager check into the dev subflake
With a couple of minor adjustments, home manager check is now moved to
the subflake. This allows to remove home-manager input from the main
flake.

Note that the CI will not (yet) find it, `nix flake check` won't see it
in the parent flake but will see it in the subflake.

Said adjustments:
- Uses config shared with the demos
- Bumped stateVersion to 25.11
- Added explicit timeout for waiting

[no ci]